\name{MetabNet-package}
\alias{MetabNet-package}
\alias{MetabNet}
\docType{package}
\title{MetabNet}
\description{
MetabNet allows targeted and untargeted network analysis
using high-resolution metabolomics data
}
\details{
\tabular{ll}{
Package: \tab MetabNet\cr
Type: \tab Package\cr
Version: \tab 0.99.1\cr
Date: \tab 2015-04-07\cr
License: \tab gpl2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Karan Uppal
Maintainer: <kuppal2@emory.edu>
}
\keyword{ MetabNet }
